+++ libcoveofds/additionalfields.py
@@ -1,4 +1,4 @@
-from libcove2.common import get_additional_fields_info, get_counts_additional_fields
+from libcove2.common import get_additional_fields_info
 from libcoveofds.schema import OFDSSchema
 
 
@@ -10,14 +10,6 @@ class AdditionalFields:
 
         schema_fields = self._schema.get_package_schema_fields()
 
-        additional_fields_all = get_additional_fields_info(json_data, schema_fields)
-
-        additional_fields = sorted(
-            get_counts_additional_fields(
-                json_data,
-                schema_fields,
-                additional_fields_info=additional_fields_all,
-            )
-        )
+        additional_fields = get_additional_fields_info(json_data, schema_fields)
 
         return additional_fields